feat(generator): auto-sync base models from bucket
